Add characterization tests for RegionViewer and friends
This adds some tests to characterize the current behavior of RegionViewer and its subcomponents.

Note that Cursor has some tests nerfed at the moment. This is becaue Cursor is currently using useRef incorrectly, in that it reads the value of said ref in the main render function, outside of an event handler or effect.