Push down react, react-dom, and styled-components dependencies
Not all packages need all of these, so we take them out of the root package.json and into the package.json's of only those packages that need them.

Note that we do get a warning from react-aria-tabpanel and react-autocomplete in the ui package, as both only have React peer dependencies up to major version 16. They do both seem to work with 17.