Allow react and react-dom 18 for peer dependencies
Most of our dependencies with a peer dep on React are now updated to a version that accepts React 18 as a peer. The one exception is the autocomplete component from react-autocomplete that underpins the Combobox and Searchbox components in the ui package. There are newer autocomplete libraries available, but these components are complex enough that writing a backwards-compatible version around a new library may be a bigger undertaking than we want to do right now.

By adding react and react-dom 18 as allowed versions in the top-level package.json, react-autocomplete no longer complains about an unmet peer dep for those. More significantly, before this change, under certain circumstances pnpm install would uninstall version 18 in favor of version 16: this change also puts a stop to that.